shipments_backend/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
psycopg2==2.9.6
isort==5.12.0
django-debug-toolbar==4.1.0
pytest==7.3.1
pytest-django==4.5.2
faker==18.10.1
django-cors-headers==4.0.0
